Stub map sidebar interactions
