+++ frontend/src/components/Map/Map.js
@@ -1,4 +1,6 @@
 import React, { useEffect, useRef, useState } from "react";
+import ReactDOMServer from "react-dom/server";
+import Popup from "./Popup";
 import "leaflet/dist/leaflet.css";
 import L from "leaflet";
 import "leaflet.markercluster/dist/leaflet.markercluster.js";
@@ -6,8 +8,6 @@ import "leaflet.markercluster/dist/MarkerCluster.css";
 import "leaflet.markercluster/dist/MarkerCluster.Default.css";
 import icon from "leaflet/dist/images/marker-icon.png";
 import iconShadow from "leaflet/dist/images/marker-shadow.png";
-import { filterActivitiesByPeriod } from "../../utils/filterActivitiesByPeriod";
-import { filterActivitiesByTime } from "../../utils/filterActivitiesByTime";
 
 let DefaultIcon = L.icon({
   iconUrl: icon,
@@ -24,13 +24,10 @@ const style = {
 function Map(props) {
   // create the basic map structure
   const mapRef = useRef(null);
-  const [FilteredActivities, setFilteredActivities] = useState(
-    props.activities
-  );
 
   useEffect(() => {
     mapRef.current = L.map("map", {
-      center: [27.6452, 30.896558],
+      center: [24.084918, 32.88612],
       zoom: 13,
       scrollWheelZoom: false,
       zoomControl: false,
@@ -48,50 +45,30 @@ function Map(props) {
 
   useEffect(() => {
     layerRef.current = L.markerClusterGroup().addTo(mapRef.current);
+    new L.Control.Zoom({ position: "topright" }).addTo(mapRef.current);
   }, []);
-
-  // add markers to layer filtered by periods selected
-  useEffect(() => {
-    
-    // filter list of activities based on selections from period facet
-    let activitiesFilteredByPeriod = [];
-    activitiesFilteredByPeriod = filterActivitiesByPeriod(
-      props.activities,
-      props.SelectedPeriod,
-      activitiesFilteredByPeriod
-    );
-
-    // filter list of activities based on selections from time slider & inputs
-    let activitiesFilteredByTime = [];
-    activitiesFilteredByTime = filterActivitiesByTime(
-      props.activities,
-      props.SelectedMinTime,
-      props.SelectedMaxTime,
-      activitiesFilteredByTime
-    );
-
-    // join filtered activities into array with only unique activties
-    let joinedActivities = [
-      ...new Set([...activitiesFilteredByPeriod, ...activitiesFilteredByTime]),
-    ];
-    setFilteredActivities(joinedActivities);
-  }, [
-    props.activities,
-    props.SelectedPeriod,
-    props.SelectedMinTime,
-    props.SelectedMaxTime,
-  ]);
 
   useEffect(() => {
     layerRef.current.clearLayers();
-    // add places to map based on filtered activities
-    FilteredActivities.forEach((activity) => {
-      const latitude = parseFloat(activity.associatedPlace[0].lat);
-      const longitude = parseFloat(activity.associatedPlace[0].lon);
+    // add places to cluster layer
+    props.Places.forEach((place) => {
+      const latitude = parseFloat(place.lat);
+      const longitude = parseFloat(place.lon);
       const latlng = { lat: latitude, lng: longitude };
-      const title = activity.associatedPlace[0].name_eng;
+      const title = place.name_eng;
 
-      L.marker(latlng, { title: title }).addTo(layerRef.current);
+      const marker = L.marker(latlng, { title: title, id: place.id });
+      marker.addTo(layerRef.current);
+      // TODO - onclick open sidebar and filter activities/articles/etc appropriately
+      marker.on("click", function (e) {
+        // test portal instead to see if that fixes the delay/sync issue with the popup responsiveness
+        this.bindPopup(
+          ReactDOMServer.renderToString(<Popup placeName={title} />)
+        );
+        props.setSelectedPlace(marker.options.id);
+        var placeSidebar = document.getElementById("placeSidebar");
+        placeSidebar.style.display = "block";
+      });
     });
   });
