--- frontend/src/components/Map/MapContainer2.css
@@ -1,5 +1,0 @@
-.loadingMap {
-  width: 100%;
-  height: 800px;
-  background-color: lightgrey;
-}
