+++ frontend/src/components/MapSidebar/ChipsSidebar.css
@@ -1,0 +1,5 @@
+.chipsNav {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
