+++ frontend/src/components/MapSidebar/ChipsSidebar.js
@@ -1,10 +1,16 @@
 import React, { useState, useEffect } from "react";
-import "./ChipsSidebar.css";
+import BackButton from "../../elements/Buttons/BackButton";
+import ResetButton from "../../elements/Buttons/ResetButton";
 
+import "./ChipsSidebar.css";
 
 function ChipsSidebar(props) {
   return (
     <div>
+      <span className="chipsNav">
+        <BackButton />
+        <ResetButton />
+      </span>
       <p>I'm a chip!</p>
     </div>
   );
