+++ frontend/src/elements/Buttons/ApplyButton.js
@@ -2,7 +2,17 @@ import React from "react";
 import "./Buttons.css";
 
 function ApplyButton() {
-  return <button className="ueeButton">Apply</button>;
+  const handleApply = () => {
+    var filterSidebar = document.getElementById("filterSidebar");
+    filterSidebar.style.display = "none";
+    var chipsSidebar = document.getElementById("chipsSidebar");
+    chipsSidebar.style.display = "block";
+  };
+  return (
+    <button className="ueeButton" onClick={handleApply}>
+      Apply
+    </button>
+  );
 }
 
 export default ApplyButton;
